+++ app/src/main/java/com/example/myapplication/adapters/HistoryAdapter.kt
@@ -1,4 +1,4 @@
-package com.example.myapplication.view
+package com.example.myapplication.adapters
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
@@ -16,7 +16,7 @@ class HistoryAdapter: RecyclerView.Adapter<HistoryAdapter.HistoryViewHolder>() {
         notifyDataSetChanged()
     }
 
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): HistoryAdapter.HistoryViewHolder {
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): HistoryViewHolder {
         val binding = HistoryFragmentRecyclerItemBinding.inflate(
             LayoutInflater.from(parent.context), parent, false
         )
