+++ app/src/main/java/com/example/myapplication/view/HistoryFragment.kt
@@ -8,6 +8,7 @@ import androidx.fragment.app.Fragment
 import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProvider
 import com.example.myapplication.R
+import com.example.myapplication.adapters.HistoryAdapter
 import com.example.myapplication.viewmodel.HistoryViewModel
 import com.example.myapplication.databinding.FragmentHistoryBinding
 import com.example.myapplication.model.AppState
