--- app/src/main/java/com/example/myapplication/view/MovieLoader.kt
@@ -1,61 +1,0 @@
-package com.example.myapplication.view
-
-import android.os.Build
-import android.os.Handler
-import android.os.Looper
-import android.util.Log
-import androidx.annotation.RequiresApi
-import com.example.myapplication.model.dto.MovieDTO
-import com.google.gson.Gson
-import java.io.BufferedReader
-import java.io.InputStreamReader
-import java.net.MalformedURLException
-import java.net.URL
-import java.util.stream.Collectors
-import javax.net.ssl.HttpsURLConnection
-
-private const val API_KEY = "3d4eed70b3bf0c001506c22b79833ff1"
-private const val LANGUAGE = "en-US"
-
-class MovieLoader(private val listener: MovieLoaderListener, private var id: Int) {
-
-    @RequiresApi(Build.VERSION_CODES.N)
-    fun loadMovie() {
-        try {
-            val uri = URL("https://api.themoviedb.org/3/movie/${id}?api_key=$API_KEY&language=$LANGUAGE")
-            val handler = Handler(Looper.myLooper()!!)
-            Thread(Runnable {
-                lateinit var urlConnection: HttpsURLConnection
-                try {
-                    urlConnection = uri.openConnection() as HttpsURLConnection
-                    urlConnection.requestMethod = "GET"
-                    urlConnection.readTimeout = 10000
-                    val bufferedReader = BufferedReader(InputStreamReader(urlConnection.inputStream))
-                    val movieDTO: MovieDTO = Gson().fromJson(getLines(bufferedReader), MovieDTO::class.java)
-                    handler.post { listener.onLoaded(movieDTO) }
-                } catch (e: Exception) {
-                    Log.e("MOVIE", "Fail connection", e)
-                    e.printStackTrace()
-                    listener.onFailed(e)
-                } finally {
-                    urlConnection.disconnect()
-                }
-            }).start()
-        } catch (e: MalformedURLException) {
-            Log.e("MOVIE", "Fail URI", e)
-            e.printStackTrace()
-            listener.onFailed(e)
-        }
-    }
-
-    @RequiresApi(Build.VERSION_CODES.N)
-    private fun getLines(reader: BufferedReader): String {
-        return reader.lines().collect(Collectors.joining("\n"))
-    }
-
-
-    interface MovieLoaderListener {
-        fun onLoaded(movieDTO: MovieDTO)
-        fun onFailed(throwable: Throwable)
-    }
-}
